fix: 🐛 fix atenArange primTupleIndex op
1. change layout variable type to int64 in atenArange op
2. deal with primTupleIndex op's input is NONE

+++ ir/include/nn_nodes/aten_arange1_node.hpp
@@ -35,9 +35,9 @@ class AtenArange1Node : public NodeMixin<AtenArange1Node, NNNode> {
 
     int64_t getDtype() const { return dtype_; }
 
-    void setLayout(int layout) { layout_ = layout; }
+    void setLayout(int64_t layout) { layout_ = layout; }
 
-    int getLayout() const { return layout_; }
+    int64_t getLayout() const { return layout_; }
 
     void setDevice(std::string device) { device_ = device; }
